requested changes and updated DB ver
