+++ src/programs/menu/break-handler/break-role-events.ts
@@ -1,10 +1,12 @@
 import { GuildMember } from "discord.js";
+import { textLog } from "../../../common/moderator";
 import {
   Command,
   CommandHandler,
   DiscordEvent,
 } from "../../../event-distribution";
 import prisma from "../../../prisma";
+import { logger } from "../common";
 
 @Command({
   event: DiscordEvent.GUILD_MEMBER_UPDATE,
@@ -12,7 +14,14 @@ import prisma from "../../../prisma";
 })
 class BreakAdded implements CommandHandler<DiscordEvent.GUILD_MEMBER_UPDATE> {
   async handle(member: GuildMember): Promise<void> {
-    await prisma.usersOnBreak.create({ data: { userId: member.id } });
+    try {
+      await prisma.usersOnBreak.create({ data: { userId: member.id } });
+    } catch (e) {
+      logger.error("Failed to register user for Break role", e);
+      await textLog(
+        `I added the break role to <@${member.id}> but couldn't register him to the DB, please contact Michel or Adrian`
+      );
+    }
   }
 }
 
@@ -22,6 +31,13 @@ class BreakAdded implements CommandHandler<DiscordEvent.GUILD_MEMBER_UPDATE> {
 })
 class BreakRemove implements CommandHandler<DiscordEvent.GUILD_MEMBER_UPDATE> {
   async handle(member: GuildMember): Promise<void> {
-    await prisma.usersOnBreak.delete({ where: { userId: member.id } });
+    try {
+      await prisma.usersOnBreak.delete({ where: { userId: member.id } });
+    } catch (e) {
+      logger.error("Failed to update Break DB", e);
+      await textLog(
+        `I removed <@${member.id}> break role, but I couldn't clean up the DB please contact Michel or Adrian.`
+      );
+    }
   }
 }
